Save the window size in the database
Load the size from the database when opening the window while maintaining alignment

+++ src/Sidekick.Database/Windows/Window.cs
@@ -12,8 +12,8 @@ namespace Sidekick.Database.Windows
         [Key]
         public string Id { get; set; }
 
-        public int Height { get; set; }
+        public double Height { get; set; }
 
-        public int Width { get; set; }
+        public double Width { get; set; }
     }
 }
